use faster eager load strategy for joined relations
It seems that the MongoJoin.options() code has a typo:
the immediateload() is used, while the comment refers to joinedload().

The immediateload() causes all nested ORM objects to be loaded via
separate SDELECT queries, so basically you get N queries for N database
rows, which is very poor for the performance.

So this commit simply changes immediateload() to joinedload().