fix join-with-empty-query bug
There is an option do do a "join" statement
with a nested query, like this:
query: {
  join: {
    'relname': {
       ...some nested query parameters...
    }
  }
}

And it breaks when you leave the query empty, like this:
query: {
  join: {
    'relname': { }
  }
}

That happens because this dictionary is used in a boolean check,
so it turns out to be `false` when the dictionary is empty,
and it breaks the code.

This commit makes a proper check (comparison with `None` object),
so now the empty query doesn't break the code.