Removed _no_joindefaults
  Adding join(()) to each query take a lot of time, especially
when we have many nested queries (e.g. join with multiple queries).
Removing this join cause time reduction from 0.8 sec to 0.3 sec for
query with 3 joins. This time is spent inside SQLalchemy when applying
all this lazyload options. We could use explicit 'join': [] in queries
if we need.
  And I am not sure that it worked correct. So for the performance and
more clean code, we should remove this.