project: respect deferred() columns and treat them with respect
@vdmit11 observed that when a model has deferred() columns,
mongosql does not load the, but reports them as loaded in
`get_full_projection_tree()`.

This is because MongoProject used to assume that by default all columns
are loaded, and did nothing for an empty projection. Naive.

Now, in the presense of deferred columns, MongoSQL always generates
load options that specify exactly what to load.

Why is this important? Because a mismatch between a projection and the
load options will lead to the N+1 problem